requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0
scikit-learn>=1.0.0,<2.0.0

# Web framework
flask>=2.0.0,<3.0.0
werkzeug>=2.0.0,<3.0.0

# Database
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<2.0.0

# HTTP requests
requests>=2.25.0,<3.0.0
urllib3>=1.26.0,<2.0.0

# Web scraping
beautifulsoup4>=4.9.0,<5.0.0
selenium>=4.0.0,<5.0.0

# Data processing
lxml>=4.6.0,<5.0.0
html5lib>=1.1,<2.0.0

# Utilities
python-dotenv>=0.19.0,<2.0.0
click>=8.0.0,<9.0.0

# Social media scraping
playwright>=1.40.0,<2.0.0
fake-useragent>=1.4.0,<2.0.0
cloudscraper>=1.2.0,<2.0.0
praw>=7.0.0,<8.0.0

# Development tools
pytest>=6.0.0,<8.0.0
pytest-cov>=2.0.0,<5.0.0
black>=21.0.0,<24.0.0
flake8>=3.8.0,<7.0.0
mypy>=0.800,<2.0.0